+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-### NEXT_VERSION_TYPE=PATCH
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [2.0.1]() (12-09-2019)
+
 * Delete deprecated objects:
 1. `Interval`
 2. `Query`
 * Delete deprecated builders:
 1. `VariableBuilder`
 2. `QueryBuilder`
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [2.0.0]() (11-09-2019)
 
 * Update class `Dashboard`: